dev/oauth2_workaround/requirements.txt
google-auth
cryptography
requests
redis
